Add React lazy and Suspense for code spliting.
