+++ src/App.js
@@ -1,17 +1,19 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, lazy, Suspense } from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 
 import { GlobalStyle } from './global.styles';
-import HomePage from './pages/HomePage';
-import ShopPage from './pages/ShopPage';
-import SigninAndSignupPage from './pages/SigninAndSignupPage';
-import CheckoutPage from './pages/Checkout';
 import Header from './components/Header';
 import NotFound from './pages/NotFound';
+import Spinner from './components/Spinner';
 import { checkUserSession } from './store/user/user.actions';
 import { selectCurrentUser } from './store/user/user.selectors';
+
+const HomePage = lazy(() => import('./pages/HomePage'));
+const ShopPage = lazy(() => import('./pages/ShopPage'));
+const CheckoutPage = lazy(() => import('./pages/Checkout'));
+const SigninAndSignupPage = lazy(() => import('./pages/SigninAndSignupPage'));
 
 function App({ checkUserSession, currentUser }) {
   useEffect(() => {
@@ -23,16 +25,22 @@ function App({ checkUserSession, currentUser }) {
       <GlobalStyle />
       <Header />
       <Switch>
-        <Route exact path="/" component={HomePage} />
-        <Route path="/shop" component={ShopPage} />
-        <Route exact path="/checkout" component={CheckoutPage} />
-        <Route
-          exact
-          path="/signin"
-          render={() => {
-            return currentUser ? <Redirect to="/" /> : <SigninAndSignupPage />;
-          }}
-        />
+        <Suspense fallback={<Spinner />}>
+          <Route exact path="/" component={HomePage} />
+          <Route path="/shop" component={ShopPage} />
+          <Route exact path="/checkout" component={CheckoutPage} />
+          <Route
+            exact
+            path="/signin"
+            render={() => {
+              return currentUser ? (
+                <Redirect to="/" />
+              ) : (
+                <SigninAndSignupPage />
+              );
+            }}
+          />
+        </Suspense>
         <Route component={NotFound} />
       </Switch>
     </div>
