+++ src/components/CollectionsOverview/index.js
@@ -1,4 +1,3 @@
-import CollectionsOverview from './CollectionsOverview';
 import CollectionsOverviewContainer from './CollectionsOverview.container';
 
-export { CollectionsOverview, CollectionsOverviewContainer };
+export default CollectionsOverviewContainer;
