+++ src/components/Spinner/Spinner.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { SpinnerContainer, SpinnerOverlay } from './Spinner.styles';
+
+const Spinner = () => (
+  <SpinnerOverlay>
+    <SpinnerContainer />
+  </SpinnerOverlay>
+);
+
+export default Spinner;
