+++ src/components/WithSpinner/WithSpinner.jsx
@@ -1,14 +1,9 @@
 import React from 'react';
-import { SpinnerContainer, SpinnerOverlay } from './WithSpinner.styles';
+
+import Spinner from '../Spinner';
 
 const WithSpinner = WrappedComponent => ({ isLoading, ...otherProps }) => {
-  return isLoading ? (
-    <SpinnerOverlay>
-      <SpinnerContainer />
-    </SpinnerOverlay>
-  ) : (
-    <WrappedComponent {...otherProps} />
-  );
+  return isLoading ? <Spinner /> : <WrappedComponent {...otherProps} />;
 };
 
 export default WithSpinner;
