+++ src/pages/Collection/index.js
@@ -1,4 +1,3 @@
-import CollectionPage from './Collection';
 import CollectionPageContainer from './Collection.container';
 
-export { CollectionPage, CollectionPageContainer };
+export default CollectionPageContainer;
